inspectai_helm_eee_compare: cross-harness EEE comparability runbook + float-timestamp fix
New runbook composes four "MMLU on pythia-6.9b" EEE artifacts that a
user might believe are mutually comparable: the public HELM EEE
conversion (mmlu:us_foreign_policy, 111 instances, prefix_exact_match,
sidecar present), two local audit reproductions of the same scenario
(quasi_prefix_exact_match, sidecars present), and an InspectAI dump
of full MMLU (13937 instances, accuracy metric, no run_spec.json).
The runbook routes them through eval-audit-from-eee and prints what
the planner can / can't conclude — the explicit answer to the user's
"how do we know two evals are really comparable?" question.

Findings the README captures:
- official vs. local audits: instance-level agree@0 = 0.94 (real
  reproducibility signal); run-level 0.0 because the public side
  emits prefix_exact_match while local emits quasi_prefix_exact_match
  (HELM schema-rename drift, not a model behavior difference).
- official vs. InspectAI: planner emits comparability_unknown:
  same_deployment, missing_run_spec, missing_scenario_class warnings
  and agree@0 collapses to None because the metric vocabularies
  don't overlap. Today's planner does signal cross-harness
  incomparability — but only indirectly via warnings + silent None.
- The InspectAI artifact carries EEE-native fields (samples_number,
  metric_config.evaluation_description, eval_library.name, …) that
  would let the planner derive same_dataset_scope / same_metric_family
  / same_eval_library facts directly from EEE without a
  run_spec.json. README documents this as the constructive followup.

Bug fix: eval_audit/normalized/loaders.py used int() on
retrieved_timestamp when sorting candidate aggregates. HELM's EEE
converter emits integer timestamps; InspectAI's emits floats
(e.g. '1777497047.279126'), which crashed core_metrics with a
ValueError for any bundle containing an InspectAI artifact. Switched
to float() — the schema treats the field as a string, so defensive
parsing is the right call.

+++ eval_audit/normalized/loaders.py
@@ -146,7 +146,7 @@ class EeeArtifactLoader(Loader):
 
         # Pick the newest by retrieved_timestamp when multiple candidates remain.
         candidates.sort(
-            key=lambda lp: int(lp[0].retrieved_timestamp or 0),
+            key=lambda lp: float(lp[0].retrieved_timestamp or 0),
             reverse=True,
         )
         chosen_log, chosen_path = candidates[0]
